consolidated staking type 3 and 4
In our articles (and in general discussion) we only mention 3 types of staking. I think even that is a lot for people to digest. 3 & 4 are clearly able to be condensed, imo. Simpler the better. 

also added that user staking is for 90 days, and that they also receive the slashing / spin up costs.
and that the protocol fees are currently 2 and 4 bps.