Matt edit 1 (#100)
* fix newsletter link

'embed' in URL means people couldn't preview it when clicked. removed now to normal.

* add Quadriga hack

* change 'crypto hacks' title to include exchanges

* small edit 

hackers more incentivized as industry and value grows

* update tps figures

2025 and 16400

* context for throughput

added old versions tps of 2-3 for context, and how DEX performance is now matched to custodial exchanges.

* update token supply

80%

* consolidated staking type 3 and 4 

In our articles (and in general discussion) we only mention 3 types of staking. I think even that is a lot for people to digest. 3 & 4 are clearly able to be condensed, imo. Simpler the better. 

also added that user staking is for 90 days, and that they also receive the slashing / spin up costs.
and that the protocol fees are currently 2 and 4 bps.

Co-authored-by: mfinestone <36642873+mfinestone@users.noreply.github.com>